Change logic of how missing data is counted
Spain wasn't getting asylum seekers from Iraq during the end of 2015, and thus incorrectly calculated as missing. Change the logic so that we check if a country has receives asylum applications from Syria, Iraq or Ukraine, and count the country as not having missing data if we find at least one of these countries.